Moved example to separate project module. Added ant build.xml to create jar files with only the necessary api files.
