+++ countly-android-example/project.properties
@@ -11,5 +11,5 @@
 #proguard.config=${sdk.dir}\tools\proguard\proguard-android.txt:proguard-project.txt
 
 # Project target.
-target=android-15
-android.library=true
+target=android-7
+android.library=false
