+++ countly-android-example/res/drawable-hdpi/ic_launcher.png
