+++ countly-android-example/res/drawable-ldpi/ic_launcher.png
