+++ countly-android-example/res/drawable-mdpi/ic_launcher.png
