+++ countly-android-example/res/drawable-xhdpi/ic_launcher.png
