+++ countly-android-example/res/layout/main.xml
