+++ countly-android-example/res/values/strings.xml
