+++ countly-sdk-android/src/org/OpenUDID/OpenUDID_manager.java
